Merge feature/calculate_ttc_in_controller into develop

+++ inc/ttc_control.h
@@ -7,9 +7,9 @@
 #include <stdbool.h>
 
 // Function prototypes
-float accel_calc(float spd);
-float ttc_calc(float dist, float spd);
+double accel_calc(double spd);
+double ttc_calc(double dist, double spd);
 void aeb_control(bool *enable_aeb, bool *alarm_cluster, bool *enable_breaking,
-                 bool *lk_seatbelt, bool *lk_doors, float *spd, float *dist);
+                 bool *lk_seatbelt, bool *lk_doors, double *spd, double *dist);
 
 #endif
